app/requirements.txt
fastapi==0.73.0
uvicorn==0.16.0
sqlalchemy==1.3.24
psycopg2-binary~=2.8
alembic==1.6.5
python-dotenv==0.19.2
black==22.3.0
yapf==0.32.0
passlib==1.7.4
fastapi_sqlalchemy==0.1.6
python-multipart==0.0.5
aiofiles==0.7.0
python-jose==3.2.0